requirement.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
sqlalchemy
pymysql
pandas
numpy
apscheduler
